+++ driver/web/rest/admin_apis.go
@@ -373,12 +373,6 @@ func (h *AdminApisHandler) CreateGroup(clientID string, current *model.User, w h
 		return
 	}
 
-	if requestData.AuthmanEnabled && !current.HasPermission("managed_group_admin") {
-		log.Printf("Only managed_group_admin could create a managed group")
-		http.Error(w, utils.NewForbiddenError().JSONErrorString(), http.StatusForbidden)
-		return
-	}
-
 	if requestData.ResearchGroup && !current.HasPermission("research_group_admin") {
 		log.Printf("'%s' is not allowed to create research group '%s'. Only user with research_group_admin permission can create research group", current.Email, requestData.Title)
 		http.Error(w, utils.NewForbiddenError().JSONErrorString(), http.StatusForbidden)
@@ -490,11 +484,6 @@ func (h *AdminApisHandler) UpdateGroup(clientID string, current *model.User, w h
 	if current != nil && !current.IsGroupsBBAdministrator() {
 		if group.CurrentMember == nil || !group.CurrentMember.IsAdmin() {
 			log.Printf("%s is not allowed to update group settings '%s'. Only group admin or all_admin_groups permission could update a group", current.Email, group.Title)
-			http.Error(w, utils.NewForbiddenError().JSONErrorString(), http.StatusForbidden)
-			return
-		}
-		if (requestData.AuthmanEnabled || group.AuthmanEnabled) && !current.HasPermission("managed_group_admin") {
-			log.Printf("%s is not allowed to update group settings '%s'. Only group admin with managed_group_admin permission could update a managed group", current.Email, group.Title)
 			http.Error(w, utils.NewForbiddenError().JSONErrorString(), http.StatusForbidden)
 			return
 		}
@@ -912,6 +901,16 @@ func (h *AdminApisHandler) DeleteGroup(clientID string, current *model.User, w h
 		return
 	}
 
+	inactive := false
+	inactiveParam, ok := r.URL.Query()["inactive"]
+	if ok && len(inactiveParam[0]) > 0 {
+		val, err := strconv.ParseBool(inactiveParam[0])
+		if err != nil {
+			log.Printf("Invalid value for inactive parameter: %s", err.Error())
+		}
+		inactive = val
+	}
+
 	group, err := h.app.Services.GetGroupEntity(clientID, id)
 	if err != nil {
 		log.Println(err.Error())
@@ -924,7 +923,7 @@ func (h *AdminApisHandler) DeleteGroup(clientID string, current *model.User, w h
 		return
 	}
 
-	err = h.app.Services.DeleteGroup(clientID, current, id)
+	err = h.app.Admin.DeleteGroup(clientID, current, id, inactive)
 	if err != nil {
 		log.Printf("Error on deleting group - %s\n", err)
 		http.Error(w, http.StatusText(http.StatusInternalServerError), http.StatusInternalServerError)